refactor(eventsub): restructure and modularize EventSub listener by modularizing
