+++ src/clients/twitch/event-sub/events/channel-follow.js
@@ -1,4 +1,4 @@
-const { replaceMessagePlaceholders } = require("./message-helpers");
+const { replaceMessagePlaceholders } = require("../utils");
 
 module.exports = async function handleChannelFollow(event) {
   try {
@@ -47,3 +47,4 @@ module.exports = async function handleChannelFollow(event) {
     );
   }
 };
+
