+++ src/clients/twitch/event-sub/events/stream-online.js
@@ -1,12 +1,19 @@
-module.exports = async function handleStreamOffline(event, liveChannels) {
+module.exports = async function handleStreamOnline(event, liveChannels) {
   try {
     const broadcasterId = event.broadcasterId;
     const broadcasterName = event.broadcasterDisplayName;
     const broadcasterLogin =
       event.broadcasterUserLogin || broadcasterName.toLowerCase();
+    const startedAt = event.startedAt ? new Date(event.startedAt) : new Date();
 
-    // Remove live status
-    liveChannels.delete(broadcasterId);
+    // Store live status
+    liveChannels.set(broadcasterId, {
+      channelId: broadcasterId,
+      channelName: broadcasterLogin,
+      displayName: broadcasterName,
+      isLive: true,
+      startedAt: startedAt,
+    });
 
     // Check if offlineOnly is enabled and bot is not paused
     const channelConfig = await fb.db.get("config", {
@@ -21,15 +28,14 @@ module.exports = async function handleStreamOffline(event, liveChannels) {
       );
       fb.log.send(
         broadcasterLogin,
-        `Live terminou! Bot reativado e funcionando normalmente ${happyEmote}`
+        `🔴 Live começando! Bot pausado para não interferir. Boa live! ${happyEmote}`
       );
     }
 
-    fb.discord.importantLog(`* ${broadcasterName} went offline`);
+    fb.discord.importantLog(`* ${broadcasterName} went live`);
   } catch (error) {
     console.log(error);
-    fb.discord.logError(
-      `Error handling stream offline event: ${error.message}`
-    );
+    fb.discord.logError(`Error handling stream online event: ${error.message}`);
   }
 };
+
