+++ src/clients/twitch/event-sub/events/vip-add.js
@@ -1,26 +1,25 @@
-module.exports = async function handleModeratorRemove(event) {
+module.exports = async function handleVipAdd(event) {
   try {
     const broadcasterId = event.broadcasterId;
     const broadcasterName = event.broadcasterDisplayName;
     const userId = event.userId;
     const botUserId = process.env.BOT_USERID;
 
-    // Only update if the bot was removed as moderator
+    // Only update if the bot was added as VIP
     if (userId === botUserId) {
       await fb.db.update(
         "config",
         { channelId: broadcasterId },
-        { $set: { botIsMod: false } }
+        { $set: { botIsVip: true } }
       );
 
       fb.discord.importantLog(
-        `* Bot was removed as moderator in channel ${broadcasterName} (${broadcasterId})`
+        `* Bot was added as VIP in channel ${broadcasterName} (${broadcasterId})`
       );
     }
   } catch (error) {
     console.log(error);
-    fb.discord.logError(
-      `Error handling moderator remove event: ${error.message}`
-    );
+    fb.discord.logError(`Error handling VIP add event: ${error.message}`);
   }
 };
+
